Clean up `ShardManager`/`ShardQueuer`/`ShardRunner` (#2653)

+++ src/model/guild/mod.rs
@@ -2244,7 +2244,7 @@ impl Guild {
     /// [`utils::shard_id`]: crate::utils::shard_id
     #[cfg(all(feature = "cache", feature = "utils"))]
     #[inline]
-    pub fn shard_id(&self, cache: impl AsRef<Cache>) -> u32 {
+    pub fn shard_id(&self, cache: impl AsRef<Cache>) -> u16 {
         self.id.shard_id(&cache)
     }
 
@@ -2269,7 +2269,7 @@ impl Guild {
     #[cfg(all(feature = "utils", not(feature = "cache")))]
     #[inline]
     #[must_use]
-    pub fn shard_id(&self, shard_count: u32) -> u32 {
+    pub fn shard_id(&self, shard_count: u16) -> u16 {
         self.id.shard_id(shard_count)
     }
